GC2D/Talk2D2 eraseBoardWindow: 95.82% -> 96.08% (preserve alpha field width)

+++ src/GC2D/Talk2D2.cpp
@@ -906,9 +906,9 @@ void TTalk2D2::setupBoardTextBox(const void* data, JMSMesgEntry* entry)
 bool TTalk2D2::eraseBoardWindow()
 {
 	bool result = false;
-	s16 alpha   = unk18->mAlpha - 4;
+	u16 alpha   = unk18->mAlpha - 4;
 
-	if (alpha < 0) {
+	if ((s16)alpha < 0) {
 		alpha                  = 0;
 		TMessageLoader* loader = unk260;
 		void* data             = loader->unk4;
